feat: settings tab
