+++ src/client/tabs/settings/AdvancedSettings.tsx
@@ -5,16 +5,10 @@ import Input from '@mui/joy/Input';
 import IconButton from '@mui/joy/IconButton';
 import FolderSpecialRounded from '@mui/icons-material/FolderSpecialRounded';
 import FormHelperText from '@mui/joy/FormHelperText';
-import SettingsRounded from '@mui/icons-material/SettingsRounded';
-import Accordion from '@mui/joy/Accordion';
-import AccordionSummary from '@mui/joy/AccordionSummary';
-import ListItemContent from '@mui/joy/ListItemContent';
-import Typography from '@mui/joy/Typography';
-import AccordionDetails from '@mui/joy/AccordionDetails';
 import React from 'react';
 import Switch from '@mui/joy/Switch';
-import { useAppDispatch, useAppSelector } from '../store/hooks';
-import { getDraftSettings, getDraftTesting, setDraftSystemDirectory, setDraftSystemTestAll, setDraftSystemTesting } from '../store/slices/system';
+import { useAppDispatch, useAppSelector } from '../../store/hooks';
+import { getDraftSettings, getDraftTesting, setDraftSystemDirectory, setDraftSystemTestAll, setDraftSystemTesting } from '../../store/slices/system';
 
 /** ------------------------------------------------------------------------- */
 
@@ -61,37 +55,23 @@ function AdvancedSettings() {
   }, [dispatch]);
 
   return (
-    <Accordion>
-      <AccordionSummary variant="soft">
-        <SettingsRounded />
-        <ListItemContent>
-          <Typography level="title-lg">Advanced</Typography>
-        </ListItemContent>
-      </AccordionSummary>
-      <AccordionDetails>
-        <Stack spacing={4} pt={1}>
-          <Stack spacing={2}>
-            <BasicTargetSettings />
-          </Stack>
-          <Stack spacing={2}>
-            <FormControl>
-              <Stack direction="row" alignItems="center" justifyContent="space-between">
-                <FormLabel>Run Discrepancy Report</FormLabel>
-                <Switch checked={enabled} onChange={handleTesting} />
-              </Stack>
-              <FormHelperText>If selected, the system will scrutinize its output against all rebate files in the "truth" folder and note any differences.</FormHelperText>
-            </FormControl>
-            <FormControl>
-              <Stack direction="row" alignItems="center" justifyContent="space-between">
-                <FormLabel>Compare All Suppliers</FormLabel>
-                <Switch disabled={!enabled} checked={compare_all} onChange={handleTestAll} />
-              </Stack>
-              <FormHelperText>If selected, discrepancies will be checked in all suppliers, not only those produced by the transformers.</FormHelperText>
-            </FormControl>
-          </Stack>
+    <Stack spacing={4} pt={1}>
+      <BasicTargetSettings />
+      <FormControl>
+        <Stack direction="row" alignItems="center" justifyContent="space-between">
+          <FormLabel>Run Discrepancy Report</FormLabel>
+          <Switch checked={enabled} onChange={handleTesting} />
         </Stack>
-      </AccordionDetails>
-    </Accordion>
+        <FormHelperText>If selected, the system will scrutinize its output against all rebate files in the "truth" folder and note any differences.</FormHelperText>
+      </FormControl>
+      <FormControl>
+        <Stack direction="row" alignItems="center" justifyContent="space-between">
+          <FormLabel>Compare All Suppliers</FormLabel>
+          <Switch disabled={!enabled} checked={compare_all} onChange={handleTestAll} />
+        </Stack>
+        <FormHelperText>If selected, discrepancies will be checked in all suppliers, not only those produced by the transformers.</FormHelperText>
+      </FormControl>
+    </Stack>
   );
 }
